% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{as.data.frame.simulation}
\alias{as.data.frame.simulation}
\title{Coerce a simulation to a data frame}
\usage{
\method{as.data.frame}{simulation}(x)
}
\arguments{
\item{x}{An object of class \code{simulation}.}
}
\value{
A data frame with columns indicating the seat allocation
for each party at each circunscription and each simulation.
}
\description{
Coerce a simulation to a data frame
}
