% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{dhondt}
\alias{dhondt}
\title{Allocate seats to votes using D'Hondt formula}
\usage{
dhondt(seats, votes)
}
\arguments{
\item{seats}{An integer with the total number of seats to allocate.}

\item{votes}{A vector with the votes (integer or proportion) per
parties.}
}
\value{
A vector of integers of the same as length \code{votes}
with the number of seats corresponding to each party
}
\description{
Allocate seats to votes using D'Hondt formula
}
\note{
The function handles ties by randomly selecting any of the
parties with ties at a given step. As a result, the function may
produce different results from the official ones.
}
