% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{share2seats}
\alias{share2seats}
\title{Allocate seats to parties according to their shares}
\usage{
share2seats(total_votes, shares, seats, threshold = 0.03)
}
\arguments{
\item{total_votes}{An integer with the total number of votes}

\item{shares}{A vector with the vote share for each shares2seats}

\item{threshold}{The proportion (between 0 and 1) of votes needed
to be allocated a seat}
}
\value{
A vector of length \code{length(shares)} with the number of
seats allocated to each party
}
\description{
Allocate seats to parties according to their shares
}
