% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{simulate_vote_share}
\alias{simulate_vote_share}
\title{Simulate vote shares for different parties}
\usage{
simulate_vote_share(p, N, ...)
}
\arguments{
\item{p}{A vector of proportions (between 0 and 1) with the vote
share for each party}

\item{N}{An integer with the number of simulations}
}
\value{
A matrix of dimensions \code{length(p)} x N with the
simulated vote shares
}
\description{
Simulates vote shares extracted from a Dirichlet distribution with
parameters given by the vote share of each party
}
