% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{dhondt}
\alias{dhondt}
\title{Allocate seats to votes using D'Hondt formula}
\usage{
dhondt(votes, seats, threshold = 0.03)
}
\arguments{
\item{votes}{A vector of total votes}

\item{seats}{An integer with the total number of seats to allocate}
}
\value{
A vector of integers of the same as length \code{votes}
with the number of seats corresponding to each party
}
\description{
Allocate seats to votes using D'Hondt formula
}
\note{
The function does not handle ties and relies on the internal
\code{order} function
}
