% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes2seats.R
\name{simulate}
\alias{simulate}
\title{Simulate seat allocations}
\usage{
simulate(total_votes, seats, p, N = 1000, threshold = 0.03, ...)
}
\arguments{
\item{total_votes}{An integer with the total number of votes}

\item{seats}{An integer with the total number of seats to be
allocated}

\item{p}{A vector of proportions (between 0 and 1) with the vote
share for each party}

\item{N}{An integer with the number of simulations. Default is
1,000}

\item{threshold}{A proportion (between 0 and 1) with the threshold}
}
\value{
A list with two elements: \code{shares} is a matrix of size
\code{length(p)} x \code{N} with the simulated vote share, and
\code{seats} of the same dimensions with the simulated seat
distribution
}
\description{
Simulate seat allocations
}
